Pull all code out into a non-static class
